util/decrypt-token/requirements.txt
pyinstaller
pypiwin32
pyaes.zip
